Added responsive design and refactoring
I've added a controller with buttons (they aren't functional yet...). When the screen size is the size of a tablet, Ipad, or anything less the controller shows up, otherwise it isn't there. I've added a logo as well as making everything fully responsive and refactored some code and added a few new things. Mario isn't visible at the moment but don't worry.